+++ R/define_S4_generic_RCM.R
@@ -193,6 +193,9 @@
 #' \item{`pbc_recdev`}{Vector of length nyears. Proportion of the bias correction to apply annually to the recruitment deviations (if estimated). 
 #' The bias correction from logspace to normal space is `exp(log_rec_dev[y] - 0.5 * pbc_recdev[y] * sigmaR^2)`. Default proportion is 1.}
 #' }
+#' \item{`pbc_earlyrecdev`}{Vector of length maxage. Proportion of the bias correction to apply to the abundance deviations in the first year of the model (if estimated). 
+#' The bias correction from logspace to normal space is `exp(log_early_rec_dev[a] - 0.5 * pbc_recdev[a] * sigmaR^2)`. Default proportion is 1.}
+#' }
 #' 
 #' @section start:
 #' Starting values can be specified in a named list for the following: